requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
pylint
black
pytest
pytest-cov
python-dotenv==0.20.0
jupyterlab
cfbd==4.4.11
